LayoutTests/http/wpt/site-isolation/overlapping-navigations-and-traversals/cross-document-traversal-cross-document-traversal.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true UseUIProcessForBackForwardItemLoading=true ] -->
<!DOCTYPE html>
<meta charset="utf-8">
<title>Cross-document traversals during cross-document traversals</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
<script type="module">
import { createIframe, waitForLoad, delay, waitForPotentialNetworkLoads } from "./resources/helpers.mjs";

promise_test(async t => {
  const iframe = await createIframe(t);

  iframe.contentWindow.location.search = "?1";
  await waitForLoad(iframe);
  await delay(t, 0);
  iframe.contentWindow.location.search = "?2";
  await waitForLoad(iframe);
  await delay(t, 0);
  iframe.contentWindow.location.search = "?3";
  await waitForLoad(iframe);
  await delay(t, 0);
  iframe.contentWindow.history.back();
  await waitForLoad(iframe);
  await delay(t, 0);
  assert_equals(iframe.contentWindow.location.search, "?2", "we made our way to ?2 for setup");

  iframe.contentWindow.history.back();
  assert_equals(iframe.contentWindow.location.search, "?2", "must not go back synchronously");

  iframe.contentWindow.history.forward();
  assert_equals(iframe.contentWindow.location.search, "?2", "must not go forward synchronously");

  iframe.onload = t.unreached_func("second load event");

  await waitForPotentialNetworkLoads(t);
  assert_equals(iframe.contentWindow.location.search, "?2", "must stay on ?2");
}, "cross-document traversals in opposite directions: the result is going nowhere");

promise_test(async t => {
  const iframe = await createIframe(t);

  iframe.contentWindow.location.search = "?1";
  await waitForLoad(iframe);
  await delay(t, 0);
  iframe.contentWindow.location.search = "?2";
  await waitForLoad(iframe);
  await delay(t, 0);

  iframe.contentWindow.history.back();
  assert_equals(iframe.contentWindow.location.search, "?2", "must not go back synchronously");

  iframe.contentWindow.history.forward();
  assert_equals(iframe.contentWindow.location.search, "?2", "must not go forward synchronously");

  iframe.onload = t.unreached_func("second load event");

  await waitForPotentialNetworkLoads(t);
  assert_equals(iframe.contentWindow.location.search, "?2", "must stay on ?2");
}, "cross-document traversals in opposite directions, second traversal invalid at queuing time but valid at the time it is run: the result is going nowhere");

promise_test(async t => {
  const iframe = await createIframe(t);

  iframe.contentWindow.location.search = "?1";
  await waitForLoad(iframe);
  await delay(t, 0);
  iframe.contentWindow.location.search = "?2";
  await waitForLoad(iframe);
  await delay(t, 0);
  iframe.contentWindow.location.search = "?3";
  await waitForLoad(iframe);
  await delay(t, 0);

  iframe.contentWindow.history.back();
  assert_equals(iframe.contentWindow.location.search, "?3", "must not go back synchronously (1)");

  iframe.contentWindow.history.back();
  assert_equals(iframe.contentWindow.location.search, "?3", "must not go back synchronously (2)");

  await waitForLoad(iframe);
  assert_equals(iframe.contentWindow.location.search, "?1", "first load event must be going back");

  iframe.onload = t.unreached_func("second load event");

  await waitForPotentialNetworkLoads(t);
  assert_equals(iframe.contentWindow.location.search, "?1", "must stay on ?1");
}, "cross-document traversals in the same (back) direction: the result is going -2 with only one load event");

promise_test(async t => {
  const iframe = await createIframe(t);

  iframe.contentWindow.location.search = "?1";
  await waitForLoad(iframe);
  await delay(t, 0);
  iframe.contentWindow.location.search = "?2";
  await waitForLoad(iframe);
  await delay(t, 0);
  iframe.contentWindow.location.search = "?3";
  await waitForLoad(iframe);
  await delay(t, 0);
  iframe.contentWindow.history.back();
  await waitForLoad(iframe);
  await delay(t, 0);
  assert_equals(iframe.contentWindow.location.search, "?2", "we made our way to ?2 for setup");
  iframe.contentWindow.history.back();
  await waitForLoad(iframe);
  await delay(t, 0);
  assert_equals(iframe.contentWindow.location.search, "?1", "we made our way to ?1 for setup");

  iframe.contentWindow.history.forward();
  assert_equals(iframe.contentWindow.location.search, "?1", "must not go forward synchronously (1)");

  iframe.contentWindow.history.forward();
  assert_equals(iframe.contentWindow.location.search, "?1", "must not go forward synchronously (2)");

  await waitForLoad(iframe);
  assert_equals(iframe.contentWindow.location.search, "?3", "first load event must be going forward");

  iframe.onload = t.unreached_func("second load event");

  await waitForPotentialNetworkLoads(t);
  assert_equals(iframe.contentWindow.location.search, "?3", "must stay on ?3");
}, "cross-document traversals in the same (forward) direction: the result is going +2 with only one load event");
</script>
